Make telststar's staging backend configurable (#104)
### Intention
While this project was only used for a single project with `peewee` as ORM there was no need to make the backend where the staged messages get stored swappable. But now I need the same library to function inside a project using `sqlalchemy`

#### Changes

There is a new `config` object in place which allows you to supply an object that manages the stages messages. Usage:

```python
from telstar import config as tlconfig
# .... 
Session = sessionmaker(bind=engine)
session = Session(autocommit=True)

tlconfig.staging.repository = StagedMessageRepository
tlconfig.staging.repository.setup(session)
``` 

We now supply a default implementation for `sqlalchemy` that behaves the same as the `peewee` implementation.  Which is verified through running the same tests with both backends. 


#### Future
I am currently not happy with how the `config` is done. What needs to improve is the handling of the `declaretive_base` because it get's quite messy to integrate that into an existing project which uses `alembic` because now you have to teach `alembic` two bases instead of just one.