Use transactional fixtures. (#132)
* Use transactional fixtures.

We make more use of the fact that `pytest.fixture` can be scope per
`session` of `function` and thus only create a single `engine` per test
run, this saves some resources and looks generally cleaner.

Also removed the difference between `realdb` and `db`. Now it's just a
single fixture that, depending on the `ORM` gives back the correct
engine.

* Fix circleci config quoting

* Transaction management is hard.

We resort back to recreating all the tables.
Not optimal but good enough for know.
The structure is likely going to change again anyway.
Will tackle this another day.