Release 1.0.0 - Delayed sending and sqlalchemy.
Why is this a major release? Simply because you will need to do a
migration when using the staging mechanism as there is a new field on
the underlying model.

And in order to not break people's apps by accident when they use and
auto update we bump this to 1.0.0 to indicate a breaking changes.