Give up control over transaction.
When running a producer in a loop with `autocommit=True` we cannot
`commit` from the repository when using the `context_callable` that
already starts a transaction.

In order to solve this we give up control over the transaction entirely
for sqlalchemy and leave it up to the user to `commit` the changes from
using `telstar.stage` - which would be the case for applications build
with frameworks already managing the transaction for the user.